chore: move hashing methods to hashingFunctions class
