+++ src/main/java/com/limechain/runtime/hostapi/functions/HashingHostFunctions.java
@@ -2,14 +2,17 @@ package com.limechain.runtime.hostapi.functions;
 
 import com.limechain.runtime.hostapi.HostApi;
 import lombok.experimental.UtilityClass;
+import net.openhft.hashing.LongHashFunction;
 import org.wasmer.ImportObject;
 import org.wasmer.Type;
 
+import java.nio.ByteBuffer;
+import java.nio.ByteOrder;
 import java.util.Arrays;
 import java.util.List;
 
 @UtilityClass
-public class HashingHostFunctions {
+public class HashingHostFunctions extends HostApi {
 
     public static List<ImportObject> getFunctions() {
         return Arrays.asList(
@@ -29,13 +32,67 @@ public class HashingHostFunctions {
                     return argv;
                 }, List.of(Type.I64), Type.I32),
                 HostFunctions.getImportObject("ext_hashing_twox_64_version_1", argv -> {
-                    return List.of(HostApi.extHashingTwox64Version1((long) argv.get(0)));
+                    return List.of(extHashingTwox64Version1((long) argv.get(0)));
                 }, List.of(Type.I64), Type.I32),
                 HostFunctions.getImportObject("ext_hashing_twox_128_version_1", argv -> {
-                    return List.of(HostApi.extHashingTwox128Version1((long) argv.get(0)));
+                    return List.of(extHashingTwox128Version1((long) argv.get(0)));
                 }, List.of(Type.I64), Type.I32),
                 HostFunctions.getImportObject("ext_hashing_twox_256_version_1", argv -> {
-                    return List.of(HostApi.extHashingTwox256Version1((long) argv.get(0)));
+                    return List.of(extHashingTwox256Version1((long) argv.get(0)));
                 }, List.of(Type.I64), Type.I32));
     }
+
+    private static int extHashingTwox64Version1(long addr) {
+        byte[] dataToHash = getDataFromMemory(addr);
+
+        byte[] hash0 = hash64(0, dataToHash);
+
+        return putDataToMemory(hash0);
+    }
+
+    private static int extHashingTwox128Version1(long addr) {
+        byte[] dataToHash = getDataFromMemory(addr);
+
+        byte[] hash0 = hash64(0, dataToHash);
+        byte[] hash1 = hash64(1, dataToHash);
+
+        ByteBuffer buffer = ByteBuffer.allocate(16);
+        buffer.put(hash0);
+        buffer.put(hash1);
+
+        byte[] byteArray = buffer.array();
+        return putDataToMemory(byteArray);
+    }
+
+    private static int extHashingTwox256Version1(long addr) {
+        byte[] dataToHash = getDataFromMemory(addr);
+
+        byte[] hash0 = hash64(0, dataToHash);
+        byte[] hash1 = hash64(1, dataToHash);
+        byte[] hash2 = hash64(2, dataToHash);
+        byte[] hash3 = hash64(3, dataToHash);
+
+        ByteBuffer buffer = ByteBuffer.allocate(32);
+        buffer.put(hash0);
+        buffer.put(hash1);
+        buffer.put(hash2);
+        buffer.put(hash3);
+
+        byte[] byteArray = buffer.array();
+        return putDataToMemory(byteArray);
+    }
+
+    private static byte[] hash64(int seed, byte[] dataToHash) {
+        final long res3 = LongHashFunction
+                .xx(seed)
+                .hashBytes(dataToHash.clone());
+
+        final ByteBuffer buffer = ByteBuffer
+                .allocate(8)
+                .order(ByteOrder.LITTLE_ENDIAN)
+                .putLong(res3);
+
+        return buffer.array();
+    }
+
 }
